Deprecates old ExperimentRegistry in favor of new ui code
